Move ExecutionProfile classes to the models package
Add Swagger annotations.
