+++ restapi/src/main/java/io/stargate/web/docsapi/service/ExecutionContext.java
@@ -19,6 +19,9 @@ import io.stargate.db.datastore.ResultSet;
 import io.stargate.db.query.BoundDMLQuery;
 import io.stargate.db.query.BoundQuery;
 import io.stargate.db.query.RowsImpacted;
+import io.stargate.web.docsapi.models.ExecutionProfile;
+import io.stargate.web.docsapi.models.ImmutableExecutionProfile;
+import io.stargate.web.docsapi.models.QueryInfo;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
